Transfer non-functioning
Going to take a look at an ATM sometime today or tomorrow to get some
ideas on how to set this up.  Have a couple ideas, but not sure how to
structure it for user friendliness.  Ideas range from having a button
that flips savings and checking to just placing a text entry field and
then two buttons that say "to savings from checking" or "to checking
from savings"